+++ findbugs-exclude.xml
@@ -19,7 +19,7 @@
 
 <FindBugsFilter>
     <Match>
-        <class name="org.wso2.extension.siddhi.io.wso2event.source.AgentCallbackImpl"/>
+        <class name="AgentCallbackImpl"/>
         <Bug pattern="WA_AWAIT_NOT_IN_LOOP"/>
     </Match>
 </FindBugsFilter>
