+++ pom.xml
@@ -26,9 +26,9 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
 
-    <groupId>org.wso2.extension.siddhi.io.wso2event</groupId>
+    <groupId>io.siddhi.extension.io.wso2event</groupId>
     <artifactId>siddhi-io-wso2event-parent</artifactId>
-    <version>5.0.1-SNAPSHOT</version>
+    <version>5.1.0-SNAPSHOT</version>
     <name>Siddhi IO WSO2Event Extensions</name>
 
     <modules>
@@ -42,7 +42,7 @@
         <siddhi.version>5.1.2</siddhi.version>
         <siddhi.version.range>[5.0.0,6.0.0)</siddhi.version.range>
         <carbon.analytics.version>3.0.3</carbon.analytics.version>
-        <siddhi.map.wso2event.version>5.0.1</siddhi.map.wso2event.version>
+        <siddhi.map.wso2event.version>5.0.2</siddhi.map.wso2event.version>
         <carbon.analytics-common.version>6.1.27</carbon.analytics-common.version>
         <carbon.uis.version>1.0.3</carbon.uis.version>
         <carbon.kernel.version>5.2.8</carbon.kernel.version>
@@ -190,7 +190,7 @@
                 <version>${carbon.analytics-common.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.wso2event</groupId>
+                <groupId>io.siddhi.extension.map.wso2event</groupId>
                 <artifactId>siddhi-map-wso2event</artifactId>
                 <version>${siddhi.map.wso2event.version}</version>
             </dependency>
@@ -419,7 +419,7 @@
                 <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.io.wso2event</groupId>
+                <groupId>io.siddhi.extension.io.wso2event</groupId>
                 <artifactId>siddhi-io-wso2event</artifactId>
                 <version>${siddhi.io.wso2event.version}</version>
             </dependency>
@@ -451,7 +451,7 @@
                 <version>${slf4j.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.io.wso2event</groupId>
+                <groupId>io.siddhi.extension.io.wso2event</groupId>
                 <artifactId>org.wso2.extension.io.wso2event.test.distribution</artifactId>
                 <version>${siddhi.io.wso2event.version}</version>
                 <type>zip</type>
