+++ tests/osgi-tests/src/test/resources/testng.xml
@@ -17,7 +17,7 @@ limitations under the License.
 <suite name="Carbon-Security-Test-Suite">
     <test name="carbon-security-tests" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.extension.siddhi.io.wso2event.test.osgi.WSO2EventSimpleFlowTestcase"/>
+            <class name="io.siddhi.extension.io.wso2event.test.osgi.WSO2EventSimpleFlowTestcase"/>
         </classes>
     </test>
 </suite>
